+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Spectral Donkey transparency fix
 - Cantrip delay now works like in the base mod
 - Spectral Armor protection scales with player tier now
+- Spectral Armor not being transparent
 
 ## 0.4.0 - 2026-07-05